FLIX-198: Customer can not create a requets if previous request in status PAYMENT_WAITING

+++ src/main/java/com/fu/flix/entity/DiscountPercent.java
@@ -19,5 +19,5 @@ public class DiscountPercent {
 
     private Double discountPercent;
 
-    private Double maxDiscountPrice;
+    private Long maxDiscountPrice;
 }
